requirements.txt
pyyaml>=6.0
rich>=13.0
reportlab>=4.0
click>=8.0
colorama>=0.4
tabulate>=0.9
jinja2>=3.1
